v1.4.1 BREAKING CHANGE - Fixing issue #22 (Geolocation.Id is more tha…

+++ Intuit.TSheets.Examples/ExampleAppService.cs
@@ -318,7 +318,7 @@ namespace Intuit.TSheets.Examples
                 Zip = "83616",
                 LinkedObjects = new LocationLinkedObjectIds
                 {
-                    Jobcodes = new List<int> { jobcode.Id }
+                    Jobcodes = new List<long> { jobcode.Id }
                 }
             };
