+++ Intuit.TSheets/Client/Core/RestClient.cs
@@ -186,7 +186,7 @@ namespace Intuit.TSheets.Client.Core
         /// <returns>The serialized response of the API method call.</returns>
         public async Task<string> DeleteAsync(
             EndpointName endpointName,
-            IEnumerable<int> ids,
+            IEnumerable<long> ids,
             LogContext logContext,
             CancellationToken cancellationToken)
         {
