+++ Intuit.TSheets/Model/Filters/GroupFilter.cs
@@ -40,7 +40,7 @@ namespace Intuit.TSheets.Model.Filters
         [JsonConverter(typeof(EnumerableToCsvConverter))]
         [JsonSchema(JsonObjectType.String)]
         [JsonProperty("ids")]
-        public IEnumerable<int> Ids { get; set; }
+        public IEnumerable<long> Ids { get; set; }
 
         /// <summary>
         /// Gets or sets the value for filtering by active or inactive state, or both.
@@ -55,7 +55,7 @@ namespace Intuit.TSheets.Model.Filters
         [JsonConverter(typeof(EnumerableToCsvConverter))]
         [JsonSchema(JsonObjectType.String)]
         [JsonProperty("manager_ids")]
-        public IEnumerable<int> ManagerIds { get; set; }
+        public IEnumerable<long> ManagerIds { get; set; }
 
         /// <summary>
         /// Gets or sets the group names you'd like to filter on.
