+++ Intuit.TSheets/Model/Jobcode.cs
@@ -58,7 +58,7 @@ namespace Intuit.TSheets.Model
         /// </summary>
         [NoSerializeOnCreate]
         [JsonProperty("id")]
-        public int Id { get; internal set; }
+        public long Id { get; internal set; }
 
         /// <summary>
         /// Gets or sets the id of the parent of this jobcode.
@@ -67,7 +67,7 @@ namespace Intuit.TSheets.Model
         /// Value is 0 if jobcode is top-level.
         /// </remarks>
         [JsonProperty("parent_id")]
-        public int? ParentId { get; set; }
+        public long? ParentId { get; set; }
 
         /// <summary>
         /// Gets or sets the name of the jobcode.
@@ -141,7 +141,7 @@ namespace Intuit.TSheets.Model
         /// </summary>
         [NoSerializeOnWrite]
         [JsonProperty("required_customfields")]
-        public IReadOnlyList<int> RequiredCustomFields { get; internal set; }
+        public IReadOnlyList<long> RequiredCustomFields { get; internal set; }
 
         /// <summary>
         /// Gets the custom field items that are to be displayed when this jobcode is chosen for a timesheet.
@@ -152,7 +152,7 @@ namespace Intuit.TSheets.Model
         /// </remarks>
         [NoSerializeOnWrite]
         [JsonProperty("filtered_customfielditems")]
-        public IReadOnlyDictionary<string, IReadOnlyList<int>> FilteredCustomFieldItems { get; internal set; }
+        public IReadOnlyDictionary<string, IReadOnlyList<long>> FilteredCustomFieldItems { get; internal set; }
 
         /// <summary>
         /// Gets or sets the status of the jobcode.
@@ -186,13 +186,13 @@ namespace Intuit.TSheets.Model
         /// </summary>
         [NoSerializeOnWrite]
         [JsonProperty("locations")]
-        public IReadOnlyList<int> Locations { get; internal set; }
+        public IReadOnlyList<long> Locations { get; internal set; }
 
         /// <summary>
         /// Gets the id of the project.
         /// </summary>
         [JsonProperty("project_id")]
-        public int? ProjectId { get; internal set; }
+        public long? ProjectId { get; internal set; }
 
 
         /// <summary>
